@@ -1,9 +1,10 @@
 package it.unive.dais.legodroid.lib.comm;
 
-import android.support.annotation.NonNull;
+
+import androidx.annotation.NonNull;
 
 /**
- * This class implements a command for the EV3 device.
+ * This class implements a command for the GenEV3 device.
  */
 public class Command extends Packet {
 
