@@ -1,27 +1,26 @@
 package it.unive.dais.legodroid.lib.plugs;
 
-import android.support.annotation.NonNull;
-
 import java.io.IOException;
 import java.util.concurrent.Future;
 
+import androidx.annotation.NonNull;
 import it.unive.dais.legodroid.lib.EV3;
 import it.unive.dais.legodroid.lib.util.Function;
 
 /**
- * Abstract class for sensors collecting reusable functionalities.
+ * Abstract class for sensors.
  */
 public abstract class AbstractSensor extends Plug<EV3.InputPort> {
     /**
-     * This field represents the type of the direct command according to the EV3 Development Kit Documentation, e.g. {@link it.unive.dais.legodroid.lib.comm.Const#EV3_COLOR}.
+     * This field represents the type of the direct command according to the GenEV3 Development Kit Documentation, e.g. {@link it.unive.dais.legodroid.lib.comm.Const#EV3_COLOR}.
      */
     protected final int type;
 
     /**
-     * Create an instance of this class given an object of type {@link it.unive.dais.legodroid.lib.EV3.Api}, and input port and the type constant.
+     * Create an instance of this class given an object of type {@link EV3.Api}, and input port and the type constant.
      * This constructor is meant for subclasses specializing a specific sensor.
      *
-     * @param api  an object of type {@link it.unive.dais.legodroid.lib.EV3.Api}.
+     * @param api  an object of type {@link EV3.Api}.
      * @param port input port where the sensor is attached to.
      * @param type type constant, e.g. {@link it.unive.dais.legodroid.lib.comm.Const#EV3_COLOR}.
      */
